Added bgzip input support. Changed cli so shared subcommand arguments are global options.
